feat: move partition consolidator and add LocalAggregator API
